pyproject.toml
[project]
name = "online-hiring-system"
version = "0.1.0"
description = "an online system that will automatically screen through resume"
authors = [
    {name = "Your Name",email = "you@example.com"}
]
license = {text = "Apache-2.0"}
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "pandas (>=2.2.3,<3.0.0)",
    "numpy (>=2.2.5,<3.0.0)",
    "redis (>=5.2.1,<6.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "flask (>=3.1.0,<4.0.0)"
]

package-mode = false

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
